// src/usercase/customer/list/dto/list.customer.dto.ts
// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface InputListCustomerDto {}

interface CustomerDto {
    id: string;
    name: string;
    address?: {
        street: string;
        number: number;
        zip: string;
        city: string;
    };
    rewardPoints?: number;
    active?: boolean;
}

export interface OutputListCustomerDto {
    customers: CustomerDto[];
}